FAC-557: refactor : Minor renaming from DeletePicture to DeleteUserPicture in service and UseCase
